attempt to consolidate ranking update functions
I ended up being able to clean up some of the logic in one of the
functions but it turns out they're all a little different. Not sure if
that's a sign that something should be changed with the data types but
I'll leave it for now.

Note left on the functions for redundancy here:

/* TODO: figure out if there's a way to consolidate these
   NOTE:
   - for updateRankedProjectNodes() the source is rankedBulletNodes of type RankedNode[]
   - for updateRankedJobNodes() the Set<string> is the target, not the source
   - for updateRankedBulletNodes() the Set<string> selectedSkillNodes is the source
   - all have targets that are of type RankedNode[]
 */